New separated script for BMO
